New FS base class and refactor opfs-ahp to use it
