+++ packages/pglite/src/fs/nodefs.ts
@@ -1,11 +1,9 @@
 import * as fs from 'fs'
 import * as path from 'path'
-import { FilesystemBase } from './types.js'
-import { PGDATA } from './index.js'
+import { EmscriptenBuiltinFilesystem, PGDATA } from './base.js'
 import type { PostgresMod, FS } from '../postgresMod.js'
-import { dumpTar, type DumpTarCompressionOptions } from './tarUtils.js'
 
-export class NodeFS extends FilesystemBase {
+export class NodeFS extends EmscriptenBuiltinFilesystem {
   protected rootDir: string
 
   constructor(dataDir: string) {
@@ -31,15 +29,7 @@ export class NodeFS extends FilesystemBase {
     return options
   }
 
-  async dumpTar(
-    mod: FS,
-    dbname: string,
-    compression?: DumpTarCompressionOptions,
-  ) {
-    return dumpTar(mod, dbname, compression)
-  }
-
-  async close(FS: FS): Promise<void> {
+  async closeFs(FS: FS): Promise<void> {
     FS.quit()
   }
 }
