+++ packages/pglite/src/fs/opfs-ahp.ts
@@ -1,13 +1,23 @@
-import { FsError } from './types.js'
-import type {
-  FsStats,
-  State,
-  FileSystemSyncAccessHandle,
-  Node,
-  FileNode,
-  DirectoryNode,
-  WALEntry,
-} from './types.js'
+import { BaseFilesystem, ERRNO_CODES, type FsStats } from './base.js'
+import type { PostgresMod, FS } from '../postgresMod.js'
+
+export interface OpfsAhpOptions {
+  initialPoolSize?: number
+  maintainedPoolSize?: number
+  debug?: boolean
+}
+
+// TypeScript doesn't have a built-in type for FileSystemSyncAccessHandle
+export interface FileSystemSyncAccessHandle {
+  close(): void
+  flush(): void
+  getSize(): number
+  read(buffer: ArrayBuffer, options: { at: number }): number
+  truncate(newSize: number): void
+  write(buffer: ArrayBuffer, options: { at: number }): number
+}
+
+// State
 
 const STATE_FILE = 'state.txt'
 const DATA_DIR = 'data'
@@ -16,21 +26,48 @@ const INITIAL_MODE = {
   FILE: 32768,
 }
 
-export interface OpfsAhpOptions {
-  root: string
-  initialPoolSize?: number
-  maintainedPoolSize?: number
+export interface State {
+  root: DirectoryNode
+  pool: PoolFilenames
 }
 
-/**
- * An OPFS Access Handle Pool VFS that exports a Node.js-like FS interface.
- * This FS is then wrapped by an Emscripten FS interface in emscriptenFs.ts.
- */
-export class OpfsAhp {
-  readyPromise: Promise<void>
-  #ready = false
+export type PoolFilenames = Array<string>
 
-  readonly root: string
+// WAL
+
+export interface WALEntry {
+  opp: string
+  args: any[]
+}
+
+// Node tree
+
+export type NodeType = 'file' | 'directory'
+
+interface BaseNode {
+  type: NodeType
+  lastModified: number
+  mode: number
+}
+
+export interface FileNode extends BaseNode {
+  type: 'file'
+  backingFilename: string
+}
+
+export interface DirectoryNode extends BaseNode {
+  type: 'directory'
+  children: { [filename: string]: Node }
+}
+
+export type Node = FileNode | DirectoryNode
+
+/**
+ * PGlite OPFS access handle pool filesystem.
+ * Opens a pool of sync access handles and then allocates them as needed.
+ */
+export class OpfsAhpFS extends BaseFilesystem {
+  declare readonly dataDir: string
   readonly initialPoolSize: number
   readonly maintainedPoolSize: number
 
@@ -55,22 +92,44 @@ export class OpfsAhp {
 
   #unsyncedSH = new Set<FileSystemSyncAccessHandle>()
 
-  constructor({ root, initialPoolSize, maintainedPoolSize }: OpfsAhpOptions) {
-    this.root = root
-    this.initialPoolSize = initialPoolSize || 1000
-    this.maintainedPoolSize = maintainedPoolSize || 100
-    this.readyPromise = this.#init()
+  constructor(
+    dataDir: string,
+    {
+      initialPoolSize = 1000,
+      maintainedPoolSize = 100,
+      debug = false,
+    }: OpfsAhpOptions = {},
+  ) {
+    super(dataDir, { debug })
+    this.initialPoolSize = initialPoolSize
+    this.maintainedPoolSize = maintainedPoolSize
   }
 
-  static async create(options: OpfsAhpOptions) {
-    const instance = new OpfsAhp(options)
-    await instance.readyPromise
-    return instance
+  async emscriptenOpts(opts: Partial<PostgresMod>) {
+    await this.#init()
+    return super.emscriptenOpts(opts)
+  }
+
+  async syncToFs(_fs: FS, relaxedDurability = false) {
+    await this.maybeCheckpointState()
+    await this.maintainPool()
+    if (!relaxedDurability) {
+      this.flush()
+    }
+  }
+
+  async closeFs(FS: FS): Promise<void> {
+    for (const sh of this.#sh.values()) {
+      sh.close()
+    }
+    this.#stateSH.flush()
+    this.#stateSH.close()
+    FS.quit()
   }
 
   async #init() {
     this.#opfsRootAh = await navigator.storage.getDirectory()
-    this.#rootAh = await this.#resolveOpfsDirectory(this.root, {
+    this.#rootAh = await this.#resolveOpfsDirectory(this.dataDir!, {
       create: true,
     })
     this.#dataDirAh = await this.#resolveOpfsDirectory(DATA_DIR, {
@@ -177,12 +236,6 @@ export class OpfsAhp {
     await this.maintainPool(
       isNewState ? this.initialPoolSize : this.maintainedPoolSize,
     )
-
-    this.#ready = true
-  }
-
-  get ready() {
-    return this.#ready
   }
 
   async maintainPool(size?: number) {
@@ -270,14 +323,6 @@ export class OpfsAhp {
       }
     }
     this.#unsyncedSH.clear()
-  }
-
-  exit(): void {
-    for (const sh of this.#sh.values()) {
-      sh.close()
-    }
-    this.#stateSH.flush()
-    this.#stateSH.close()
   }
 
   // Filesystem API:
@@ -674,5 +719,17 @@ export class OpfsAhp {
       ah = await ah.getDirectoryHandle(part, { create: options?.create })
     }
     return ah
+  }
+}
+
+class FsError extends Error {
+  code?: number
+  constructor(code: number | keyof typeof ERRNO_CODES | null, message: string) {
+    super(message)
+    if (typeof code === 'number') {
+      this.code = code
+    } else if (typeof code === 'string') {
+      this.code = ERRNO_CODES[code]
+    }
   }
 }
