+++ packages/pglite/tsup.config.ts
@@ -19,7 +19,7 @@ const replaceAssertPlugin = {
 const entryPoints = [
   'src/index.ts',
   'src/fs/nodefs.ts',
-  'src/fs/opfs-ahp/index.ts',
+  'src/fs/opfs-ahp.ts',
   'src/templating.ts',
   'src/live/index.ts',
   'src/vector/index.ts',
